django/requirements.txt
asgiref==3.10.0
bcrypt==5.0.0
cffi==2.0.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
Deprecated==1.3.1
Django==5.2.8
django-environ==0.12.0
djangorestframework==3.16.1
fabric==3.2.2
fake==0.8
Faker==38.0.0
fonttools==4.60.1
invoke==2.2.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
paramiko==4.0.0
pillow==12.0.0
psutil==7.1.3
pycparser==2.23
PyNaCl==1.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
wrapt==2.0.1
